requirements.txt
pandas>=2.0.0
numpy>=1.24.0
ccxt>=4.0.0
PyYAML>=6.0
# Indicadores técnicos - implementaciones propias (sin dependencias externas problemáticas)
# pandas-ta>=0.4.71  # Removido por problemas de compatibilidad con Python 3.13
MetaTrader5>=5.0.45
pytest>=8.0.0
pytest-asyncio>=0.21.0
scikit-learn>=1.3.0
streamlit>=1.28.0
plotly>=5.17.0

# HTTP requests (for dashboard verification)
requests>=2.31.0

# Additional utilities
websockets>=12.0

# Development and debugging
ipykernel>=6.25.0

# Data validation
pydantic>=2.5.0

# Enhanced logging
colorlog>=6.8.0

# Visualization
seaborn>=0.13.0
matplotlib>=3.8.0
